Added CRUD Operations/More Stability
Added the ability to view, delete, and update  medication and added to general flow. Update is still finnicky but I'll work on that tomorrow. Did some refactoring to make it way more readable and will continue with that. Also going to start working on daily tracking soon then UI is next after that. And performance. Lots to do.