Adding Viewing Logs and Made Medication/Log Viewing More consistent
Added the ability to view logs. I refactored the layout of the app so
that the main screen is simpler. Viewing medications and entries are now
separate screens accessed from the main activity. Fixed a bug in my log
entry database and changed the schema of it so it can be searched
somewhat easier