Refactoring!
Not much if any changes to functionality or anything, just tidying up of functions (especially all the filtering stuff in the ViewLogListActivity) and adding of comments to most of the log entry-related classes. Next up, continue trying to make it look OK, and a daily requirements list. Maybe notifications too.